+++ src/features/colinks/LinkHolders.tsx
@@ -7,7 +7,7 @@ import { client } from '../../lib/gql/client';
 import { Flex, Text } from '../../ui';
 
 import { CoLinksNameAndAvatar } from './CoLinksNameAndAvatar';
-import { QUERY_KEY_COLINKS } from './CoLinksWizard';
+import { QUERY_KEY_COLINKS } from './wizard/CoLinksWizard';
 
 const fetchLinkHolders = async (target?: string, limit?: number) => {
   const { link_holders } = await client.query(
