+++ src/pages/colinks/ActivityPage.tsx
@@ -6,11 +6,11 @@ import { isFeatureEnabled } from '../../config/features';
 import { ActivityList } from '../../features/activities/ActivityList';
 import { CoLinksChainGate } from '../../features/colinks/CoLinksChainGate';
 import { CoLinksHistory } from '../../features/colinks/CoLinksHistory';
-import { QUERY_KEY_COLINKS } from '../../features/colinks/CoLinksWizard';
 import { Leaderboard } from '../../features/colinks/Leaderboard';
 import { PostForm } from '../../features/colinks/PostForm';
 import { RightColumnSection } from '../../features/colinks/RightColumnSection';
 import { useCoLinks } from '../../features/colinks/useCoLinks';
+import { QUERY_KEY_COLINKS } from '../../features/colinks/wizard/CoLinksWizard';
 import { CoSoulGate } from '../../features/cosoul/CoSoulGate';
 import { Clock, Star } from '../../icons/__generated';
 import { paths } from '../../routes/paths';
