+++ src/routes/routes.tsx
@@ -26,7 +26,7 @@ import {
 import { DebugLogger } from '../common-lib/log';
 import { isFeatureEnabled } from '../config/features';
 import { CoLinksLayout } from '../features/colinks/CoLinksLayout';
-import { CoLinksWizardLayout } from '../features/colinks/CoLinksWizardLayout';
+import { CoLinksWizardLayout } from '../features/colinks/wizard/CoLinksWizardLayout';
 import AddMembersPage from '../pages/AddMembersPage/AddMembersPage';
 import CircleActivityPage from '../pages/CircleActivityPage';
 import { ActivityPage } from '../pages/colinks/ActivityPage';
@@ -38,8 +38,8 @@ import { LinkHoldingsPage } from '../pages/colinks/LinkHoldingsPage';
 import { RepScorePage } from '../pages/colinks/RepScorePage';
 import { TradesPage } from '../pages/colinks/TradesPage';
 import { ViewProfilePage } from '../pages/colinks/ViewProfilePage/ViewProfilePage';
-import { WizardPage } from '../pages/colinks/WizardPage';
-import { WizardStart } from '../pages/colinks/WizardStart';
+import { WizardPage } from '../pages/colinks/wizard/WizardPage';
+import { WizardStart } from '../pages/colinks/wizard/WizardStart';
 import CoSoulExplorePage from '../pages/CoSoulExplorePage/CoSoulExplorePage';
 import GivePage from '../pages/GivePage';
 import { InviteCodePage } from '../pages/InviteCodePage';
